requirements.txt
astroid==2.2.5
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
oauthlib==3.0.1
progressbar2==3.42.0
pylint==2.3.1
PySocks==1.7.0
python-utils==2.3.0
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.1
tweepy==3.7.0
typed-ast==1.4.0
urllib3==1.25.3
wrapt==1.11.2
